refactor(state): simplified / fixed state management * now the different reducers just need to return an object with the changed state as keys, no more redudant Object.assign (which breaks things too) * related experiments & cleanups
