+++ packages/web/src/ui/flow/viewer.js
@@ -24,30 +24,30 @@ const reducers = {
       },
       autorotate: false
     }
-    return Object.assign({}, state, {viewer})
+    return {viewer}
   },
   toggleAutorotate: (state, autoRotate) => {
     const controls = Object.assign({}, state.viewer.controls, {autoRotate: {enabled: autoRotate}})
     const viewer = Object.assign({}, state.viewer, {controls})
-    return Object.assign({}, state, {viewer})
+    return {viewer}
   },
   toggleGrid: (state, show) => {
     const grid = Object.assign({}, state.viewer.grid, {show})
     const viewer = Object.assign({}, state.viewer, {grid})
-    return Object.assign({}, state, {viewer})
+    return {viewer}
   },
   toggleAxes: (state, show) => {
     const axes = Object.assign({}, state.viewer.axes, {show})
     const viewer = Object.assign({}, state.viewer, {axes})
-    return Object.assign({}, state, {viewer})
+    return {viewer}
   },
   toPresetView: (state, viewName) => {
     const viewer = Object.assign({}, state.viewer, {camera: {position: viewName}})
-    return Object.assign({}, state, {viewer})
+    return {viewer}
   },
   setProjectionType: (state, projectionType) => {
     const viewer = Object.assign({}, state.viewer, {camera: {projectionType}})
-    return Object.assign({}, state, {viewer})
+    return {viewer}
   }
 
 }
